[PROF-12377] Update process context support with latest version of reference library
**What does this PR do?**

This PR upgrades the process context support with the latest version
of the latest library, taken from
https://github.com/DataDog/fullhost-code-hotspots-wip/tree/b33673d801b85a6c38fa0e9f1a139cb246737ce8/lang-exp/anonmapping-clib

Of interest, this version of the context support library:
* Introduces more fields for publishing information
* Simplifies management of active contexts by moving the "keep a static
  with the current context" inside the library itself
* Introduces a no-op version of the library, so we no longer need a
  `otel_process_ctx_macos.cpp` on this side

On the Java side, I:
* Used the Datadog names for fields when setting them -- hopefully this
  will avoid any confusion on which fields map to what on the OTEL side
* Removed the `ReentrantReadWriteLock`, replacing it with a
  `synchronized`, as it's not safe to concurrently read and write the
  current context -- all access must be serialized.
* Removed the `verifyMappingStructure` as I thought it's a bit awkward
  to maintain if we need to change things until we stabilize the
  feature. Instead, I think for now it's simpler to assume the native
  reader is correct.

**Motivation:**

Make sure Java is up-to-date with the latest improvements on the lib!

Also, since I had a lot of context on the latest changes, seems easier
for me to shoot out a PR for this update.

**Additional Notes:**

I ran `./gradlew test` but I'm not sure the tests are running
properly... Am I missing something?

**How to test the change?**

I've updated the tests but at least with `./gradlew test` they don't
seem to be running. Hopefully in CI it's all wired up?