[PROF-12372] Update process context support with version 0.0.7 of reference library
**What does this PR do?**

This PR upgrades the process context support with the latest version
(0.0.7) of the reference library, taken from
<https://github.com/DataDog/fullhost-code-hotspots-wip/tree/5ec6258b3bd498d5aded14e081bc740a8134ed54/lang-exp/anonmapping-clib>.

Of interest, this version of the context support library:
* Fixes a build issue on the noop build (non-Linux support)
* Fixes dropping the mapping not fully dropping the mapping correctly
* Makes reading work on Linux < 5.17

No changes are needed on the Java side to adopt these improvements.

**Motivation:**

Get latest version into dd-trace-java!

**Additional Notes:**

N/A

**How to test the change?**

The current test coverage should be enough to validate these changes.